+++ CHANGELOG.md
@@ -88,5 +88,6 @@
 - Anonymize user's IP addresses before logging them outside the application, by removing the last octet of the address. Also use Rollbar's built-in IP address anonymizer.
 - BEIS users can view Transactions & Budgets on a project, but not create or edit them
 - Country list for recipient countries when creating an activity has been reduced to only those ODA uses as recipients.
+- Add feedback form link to phase banner
 
 [release-2]: https://github.com/dxw/DataSubmissionService/compare/release-2...release-1